requirements.txt
amqp==2.5.1
bcrypt==3.1.4
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.235
botocore==1.12.235
celery==4.3.0
certifi==2019.6.16
cffi==1.13.0
chardet==3.0.4
configparser==3.8.1
contextlib2==0.5.5
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-bmemcached==0.2.4
django-cors-headers==3.1.0
django-dotenv==1.4.2
django-filter==2.0.0
django-rest-auth==0.9.5
django-storages==1.7.1
django-stubs==1.2.0
djangorestframework==3.10.2
djangorestframework-stubs==0.3.0
docutils==0.15.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
jmespath==0.9.4
kombu==3.0.37
more-itertools==7.2.0
mypy==0.740
mypy-extensions==0.4.3
oauthlib==3.1.0
pathlib2==2.3.4
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.6.4
python-binary-memcached==0.28.0
python-dateutil==2.8.0
python-http-client==3.2.1
python-openid==2.2.5
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
scandir==1.10.0
sendgrid==6.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.7.4
uhashring==1.1
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
zipp==0.6.0
pillow==6.1.0
pyjwt==1.6.4
